cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crowdpad-cli"
version = "0.1.0"
description = "Create a virtual joypad and listens for commands from a remote server."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click==8.3.0",
    "pydantic==2.12.0",
    "python-dotenv==1.1.1",
    "python-uinput==1.0.1; sys_platform == 'linux'",
    "websockets==15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]

[project.scripts]
start = "crowdpad_cli.main:cli"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.ruff.format]
quote-style = "single"
